Add quicksort and helper functions
