+++ rust-algos/search-algos/src/binary_search.rs
@@ -1,0 +1,1 @@
+// TODO: Implement Binary Search
