+++ rust-algos/search-algos/src/bubble_sort.rs
@@ -1,5 +1,3 @@
-use rand::Rng;
-
 pub fn bubble_sort(vec: &mut Vec<i32>) {
     let mut left = 0;
     let n = vec.len();
@@ -31,30 +29,10 @@ pub fn optimized_bubble_sort(vec: &mut Vec<i32>) {
     }
 }
 
-pub fn check_sorted(vec: &Vec<i32>) -> bool {
-    for index in 1..vec.len() {
-        if vec[index - 1] > vec[index] {
-            return false;
-        }
-    }
-    true
-}
-
-pub fn make_random_vec(num_items: i32, max: i32) -> Vec<i32> {
-    let mut rng = rand::thread_rng();
-
-    let mut vec: Vec<i32> = Vec::with_capacity(num_items as usize);
-
-    for _ in 0..num_items {
-        vec.push(rng.gen_range(0..max));
-    }
-
-    vec
-}
-
 #[cfg(test)]
 mod test {
-    use crate::bubble_sort::{bubble_sort, check_sorted, make_random_vec, optimized_bubble_sort};
+    use crate::bubble_sort::{bubble_sort, optimized_bubble_sort};
+    use crate::helpers::{check_sorted, make_random_vec};
 
     #[test]
     fn test_random_list() {
